The name of the file is printed once it is complete.
The idea of this is that you get a better view of what units are taking what time. When printing out at the beginning, you tend to get large clumps of output all appearing at the same time with some wait leading up to that with no feedback at all.