[2] [GCC] The build may be configured to allow unused variables.
For example; so that a debug build can be more permissive around its warnings. While this could trigger more broken release builds, in theory, it has the advantage that while code is in a state of flux, you're not forced to temporarily comment out a member variable.

[2] [Clang] The build may be configured to allow unused variables.

For example; so that a debug build can be more permissive around its warnings. While this could trigger more broken release builds, in theory, it has the advantage that while code is in a state of flux, you're not forced to temporarily comment out a member variable.